Separate HWI urls to new file
