+++ src/cryptoadvance/specter/controller.py
@@ -31,9 +31,6 @@ DEBUG = True
 from flask import current_app as app
 rand = random.randint(0, 1e32) # to force style refresh
 
-from .hwi_rpc import HWIBridge
-
-
 ########## template injections #############
 @app.context_processor
 def inject_debug():
@@ -42,23 +39,6 @@ def inject_debug():
     return dict(debug=app.config['DEBUG'])
 
 ################ routes ####################
-
-hwi = HWIBridge()
-
-@app.route("/hwi_rpc/", methods=["POST"])
-@login_required
-def api():
-    """JSON-RPC for ... anything. In this case - HWI Bridge"""
-    try:
-        data = json.loads(request.data)
-    except:
-        return jsonify({
-            "jsonrpc": "2.0",
-            "error": { "code": -32700, "message": "Parse error" },
-            "id": None
-        }), 500
-    return jsonify(hwi.jsonrpc(data))
-
 @app.route('/wallets/<wallet_alias>/combine/', methods=['GET', 'POST'])
 @login_required
 def combine(wallet_alias):
@@ -612,12 +592,6 @@ def wallet_settings(wallet_alias):
 def new_device():
     app.specter.check()
     return render_template("device/new_device.jinja", specter=app.specter, rand=rand)
-
-@app.route('/hwi_new_device/')
-@login_required
-def hwi_new_device_xpubs():
-    app.specter.check()
-    return render_template("device/hwi_new_device_xpubs.jinja", specter=app.specter, rand=rand)
 
 @app.route('/new_device/<device_type>/', methods=['GET', 'POST'])
 @login_required
